fix type error

+++ src/timbuk_parser-nobison.cc
@@ -1077,7 +1077,7 @@ Automata<Symbol> TimbukParser<Symbol>::from_line_to_automaton(std::string line) 
                 // }
             }
             for (const auto &t2 : aut2.transitions) {
-                int Q = aut.qubitNum + t2.first.symbol().qubit(); // we need to shift the qubit number
+                int Q = aut.qubitNum + static_cast<int>(t2.first.symbol().qubit()); // we need to shift the qubit number
                 if (t2.first.is_internal()) { // if the to-be-appended transition is internal, then
                     for (const auto &kv : t2.second) { // for each pair of top -> ...
                         auto top = kv.first;
